Made space table as extension of base space
1. Changed Campuse module to Site module. Still APIs needs to be changed.
2. By default all spaces will have siteId, buildingId and floorId. They are populated contextually.
3. Bug fix in SelectRecordBuilder and InsertRecordBuilder
4. Addition of Zone is yet to be handled

+++ src/main/java/com/facilio/leed/util/LeedAPI.java
@@ -114,7 +114,7 @@ public class LeedAPI {
 				context.setEnergyScore(rs.getLong("HUMANEXPERIENCESCORE"));
 				context.setEnergyScore(rs.getLong("TRANSPORTSCORE"));
 				context.setEnergyScore(rs.getLong("LEEDSCORE"));
-				context.setArea(rs.getLong("AREA"));
+				context.setArea(rs.getInt("AREA"));
 				context.setName(rs.getString("NAME"));
 				context.setPhotoId(rs.getLong("PHOTO_ID"));
 			}
